requirements.txt
# Hyperspectral-Restruct Dependencies
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Data handling
pandas>=2.0.0
h5py>=3.8.0

# Hyperspectral data
spectral>=0.23.0
rasterio>=1.3.8

# API & web
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# ONNX for edge deployment
onnx>=1.14.0
onnxruntime>=1.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Geospatial (optional but recommended)
shapely>=2.0.0

# Development
pytest>=7.4.0
black>=23.7.0
ruff>=0.0.280
